moved the tests and updated changelog
